requirements.txt
torch 
torchvision 
torchaudio
transformers
sentence-transformers
faiss-cpu
datasets
fastapi 
uvicorn
moto
streamlit
scikit-learn
python-dotenv
openai
python-dotenv
pydantic_settings
streamlit-mic-recorder